Fix single key architecture (#4)
The single key architecture without time bounded key derivation was vulnerable to catastrophic loss with a single key compromise. Although a key compromise in the current configuration could still be considered catastrophic this is a big step towards more reasonably performing key rotation. I still need to work on a migration strategy where two versions of a key could be used simultaneously during a migration. The time bounded key derivation, however, mitigates much of the priority and need to consider that in the immediate future.